+++ shells.nix
@@ -38,17 +38,21 @@ in
                   # the shell (but not for dependencies of it)
                   # that is the reason we are not using the check
                   # variant of the matrix
-                  drv = enableChecks drv';
+                  drv =
+                    let
+                      checkedDrv = (enableChecks drv');
+                    in
+                    # Priority of nativeBuildInputs is: 
+                      # 1. nativeBuildInputs
+                      # 2. checkInputs
+                      # 3. shellInputs
+                      # 4. shellCommands
+                    checkedDrv // { nativeBuildInputs = checkedDrv.nativeBuildInputs or [ ] ++ drv.shellInputs or [ ] ++ (lib.optional (drv ? shellCommands) drv.shellCommands); };
                   targetName = "${component.name}.${name}";
                   shellCommandsDesc = lib.filterAttrs (_: value: value.show) (drv.shellCommands or { })._descriptions or { };
                   shellPkg = (drv.drvAttrs // {
                     inherit (drv) passthru;
                     name = "${targetName}-shell";
-
-                    # this will get merged with nativeBuildInputs
-                    # from "drv" inside mkShell so no need to
-                    # add to it here
-                    nativeBuildInputs = drv.shellInputs or [ ] ++ (lib.optional (drv ? shellCommands) drv.shellCommands);
 
                     # we will get double shellhooks if we do not
                     # remove this here