+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - shellcheck from the ci attribute will now find more shell scripts and is faster.
+- Changed order of shell dependencies so shellInputs can't change the behavior of checks in the shell.
 
 ### Changed Versions
 - nixpkgs from 21.11 to 22.05